Move wao logic into ssa
Necesarry because writes to attrs need to follow the same logic as
returns
