+++ ast_tools/passes/__init__.py
@@ -7,4 +7,3 @@ from .ssa import *
 from .util import *
 from .loop_unroll import loop_unroll
 from .if_inline import if_inline
-from .write_attrs_once import write_attrs_once
